# scalar quantization: add SIMD l2 distance calculation