… adapters/repos/db/vector/hnsw/distancer/asm/l2_float_byte_avx256.go
… adapters/repos/db/vector/hnsw/distancer/asm/l2_float_byte_avx256.s