### What's being changed:

```
benchstat sq_l2_go.txt sq_l2_avx256_inline.txt
goos: linux
goarch: amd64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
                               │ sq_l2_go.txt │       sq_l2_avx256_inline.txt        │
                               │    sec/op    │    sec/op     vs base                │
_L2FloatByte/vector_dim=1-4      0.8903n ± 0%   2.6725n ± 1%  +200.18% (p=0.002 n=6)
_L2FloatByte/vector_dim=2-4       1.622n ± 1%    2.969n ± 1%   +83.01% (p=0.002 n=6)
_L2FloatByte/vector_dim=3-4       3.333n ± 4%    3.579n ± 7%    +7.38% (p=0.002 n=6)
_L2FloatByte/vector_dim=4-4       3.648n ± 1%    3.860n ± 1%    +5.80% (p=0.002 n=6)
_L2FloatByte/vector_dim=5-4       4.276n ± 2%    5.340n ± 0%   +24.87% (p=0.002 n=6)
_L2FloatByte/vector_dim=6-4       4.336n ± 1%    5.042n ± 0%   +16.28% (p=0.002 n=6)
_L2FloatByte/vector_dim=8-4       3.560n ± 0%    5.930n ± 0%   +66.59% (p=0.002 n=6)
_L2FloatByte/vector_dim=10-4      4.149n ± 1%    7.708n ± 0%   +85.78% (p=0.002 n=6)
_L2FloatByte/vector_dim=12-4      4.745n ± 0%    8.614n ± 1%   +81.52% (p=0.002 n=6)
_L2FloatByte/vector_dim=16-4      5.924n ± 3%    7.009n ± 2%   +18.32% (p=0.002 n=6)
_L2FloatByte/vector_dim=24-4      8.313n ± 1%    7.394n ± 2%   -11.06% (p=0.002 n=6)
_L2FloatByte/vector_dim=30-4      10.09n ± 0%    10.17n ± 1%    +0.89% (p=0.002 n=6)
_L2FloatByte/vector_dim=31-4      10.40n ± 0%    10.82n ± 1%    +4.09% (p=0.002 n=6)
_L2FloatByte/vector_dim=32-4     10.690n ± 0%    7.215n ± 1%   -32.51% (p=0.002 n=6)
_L2FloatByte/vector_dim=64-4     20.210n ± 0%    9.555n ± 3%   -52.72% (p=0.002 n=6)
_L2FloatByte/vector_dim=67-4      21.08n ± 0%    11.76n ± 4%   -44.20% (p=0.002 n=6)
_L2FloatByte/vector_dim=128-4     39.19n ± 0%    15.25n ± 0%   -61.09% (p=0.002 n=6)
_L2FloatByte/vector_dim=256-4     84.15n ± 0%    28.93n ± 3%   -65.62% (p=0.002 n=6)
_L2FloatByte/vector_dim=260-4     85.28n ± 0%    28.86n ± 3%   -66.16% (p=0.002 n=6)
_L2FloatByte/vector_dim=299-4     97.05n ± 0%    31.65n ± 4%   -67.38% (p=0.002 n=6)
_L2FloatByte/vector_dim=300-4     97.25n ± 1%    31.80n ± 4%   -67.30% (p=0.002 n=6)
_L2FloatByte/vector_dim=384-4    122.30n ± 0%    37.18n ± 1%   -69.60% (p=0.002 n=6)
_L2FloatByte/vector_dim=390-4    124.15n ± 0%    37.16n ± 0%   -70.06% (p=0.002 n=6)
_L2FloatByte/vector_dim=600-4    186.15n ± 0%    51.84n ± 1%   -72.15% (p=0.002 n=6)
_L2FloatByte/vector_dim=768-4    235.85n ± 0%    60.90n ± 0%   -74.18% (p=0.002 n=6)
_L2FloatByte/vector_dim=777-4    238.05n ± 0%    61.29n ± 0%   -74.25% (p=0.002 n=6)
_L2FloatByte/vector_dim=784-4    240.85n ± 0%    62.79n ± 1%   -73.93% (p=0.002 n=6)
_L2FloatByte/vector_dim=1024-4   311.50n ± 0%    76.47n ± 1%   -75.45% (p=0.002 n=6)
_L2FloatByte/vector_dim=1536-4    464.2n ± 1%    107.1n ± 1%   -76.93% (p=0.002 n=6)
```

### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.